src/com/sanushi/junit/basics/BankAccount.java
package com.sanushi.junit.basics;

/**
 * Created by Sanushi Salgado on 28/05/2020
 */
public class BankAccount {
    private String firstName;
    private String lastName;
    private double balance;
    private int accountType;

    public static final int CHECKING = 1;
    public static final int SAVINGS = 2;

    public BankAccount(String firstName, String lastName, double balance, int accountType) {
        this.firstName = firstName;
        this.lastName = lastName;
        this.balance = balance;
        this.accountType = accountType;
    }

    // The branch argument is true, if the customer is performing the transaction at
    // Utilities branch, with Utilities teller
    // False, if the customer is performing the transaction at an ATM
    public double deposit(double amount, boolean branch) {
        balance += amount;
        return balance;
    }

    public double withdraw(double amount, boolean branch) throws IllegalAccessException {
        if(amount > 500 && !branch)
            throw new IllegalArgumentException();

        balance -= amount;
        return balance;
    }

    public double getBalance() {
        return balance;
    }

    public boolean isChecking() {
        return accountType == CHECKING;
    }

}
